@@ -43,7 +43,7 @@ class ImageColorsPage extends StatelessWidget {
       );
 
       final ImagePicker _picker = ImagePicker();
-      String path;
+      String? path;
       if (selected == 0) {
         // 调用相机
         path = (await _picker.getImage(source: ImageSource.camera))?.path;
@@ -89,7 +89,7 @@ class ImageColorsPage extends StatelessWidget {
                   Container(
                       margin: EdgeInsets.all(15),
                       child: PaletteSwatches(
-                          generator: controller.paletteGenerator.value)),
+                          generator: controller.paletteGenerator!.value)),
                 ],
               ),
             );
@@ -115,7 +115,7 @@ class PaletteSwatches extends StatelessWidget {
   ///
   /// The [generator] is optional. If it is null, then the display will
   /// just be an empty container.
-  const PaletteSwatches({Key key, @required this.generator}) : super(key: key);
+  const PaletteSwatches({Key? key, required this.generator}) : super(key: key);
 
   /// The [PaletteGenerator] that contains all of the swatches that we're going
   /// to display.
@@ -163,16 +163,16 @@ class PaletteSwatch extends StatelessWidget {
   /// If the [color] argument is omitted, then the swatch will show a
   /// placeholder instead, to indicate that there is no color.
   const PaletteSwatch({
-    Key key,
+    Key? key,
     this.color,
     this.label,
   }) : super(key: key);
 
   /// The color of the swatch. May be null.
-  final Color color;
+  final Color? color;
 
   /// The optional label to display next to the swatch.
-  final String label;
+  final String? label;
 
   @override
   Widget build(BuildContext context) {
@@ -222,7 +222,7 @@ class PaletteSwatch extends StatelessWidget {
               children: <Widget>[
                 swatch,
                 Container(width: 5.0),
-                Text(label),
+                Text(label!),
               ],
             ),
           ),
