@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
 class LoginState {
-  TextEditingController phoneNumberController;
-  TextEditingController passwordController;
+  late TextEditingController phoneNumberController;
+  late TextEditingController passwordController;
 
   final _isChecked = false.obs;
 
