@@ -1,11 +1,10 @@
 import 'package:dio/dio.dart';
-import 'package:flutter/material.dart';
 import 'package:flutter_easy/flutter_easy.dart';
 
 export 'constant.dart';
 export 'package:session/session.dart' show Result;
 
-Config _config(String baseURL) {
+Config _config(String? baseURL) {
   Config.logEnable = false;
 
   return Config(
@@ -23,8 +22,7 @@ SessionInterceptorSendHandler _onRequest = (options) async {
 };
 
 /// 响应结果拦截处理
-Result _onValidResult<T>(
-    Result result, bool validResult, BuildContext context) {
+Result _onValidResult<T>(Result result, bool validResult) {
   logResponse(result);
   return result;
 }
@@ -36,15 +34,13 @@ Result _onValidResult<T>(
 /// path: 请求路径
 /// queryParameters: URL携带请求参数
 /// validResult: 是否检验返回结果
-/// context: 上下文
 /// autoLoading: 展示Loading
 ///
 Future<Result> getHttpBin(
-    {String baseUrl,
+    {String? baseUrl,
     String path = '',
-    Map<String, dynamic> queryParameters,
+    Map<String, dynamic>? queryParameters,
     bool validResult = true,
-    BuildContext context,
     bool autoLoading = false}) async {
   return requestHttpBin(
       baseUrl: baseUrl,
@@ -52,7 +48,6 @@ Future<Result> getHttpBin(
       queryParameters: queryParameters,
       options: Options(method: 'get'),
       validResult: validResult,
-      context: context,
       autoLoading: autoLoading);
 }
 
@@ -63,15 +58,13 @@ Future<Result> getHttpBin(
 /// path: 请求路径
 /// data: 请求参数
 /// validResult: 是否检验返回结果
-/// context: 上下文
 /// autoLoading: 展示Loading
 ///
 Future<Result> postHttpBin(
-    {String baseUrl,
+    {String? baseUrl,
     String path = '',
     data,
     bool validResult = true,
-    BuildContext context,
     bool autoLoading = false}) async {
   return requestHttpBin(
       baseUrl: baseUrl,
@@ -79,7 +72,6 @@ Future<Result> postHttpBin(
       data: data,
       options: Options(method: 'post'),
       validResult: validResult,
-      context: context,
       autoLoading: autoLoading);
 }
 
@@ -91,17 +83,15 @@ Future<Result> postHttpBin(
 /// path: 请求路径
 /// data: 请求参数
 /// validResult: 是否检验返回结果
-/// context: 上下文
 /// autoLoading: 展示Loading
 ///
 Future<Result> requestHttpBin(
-    {String baseUrl,
+    {String? baseUrl,
     String path = '',
     data,
-    Map<String, dynamic> queryParameters,
-    Options options,
+    Map<String, dynamic>? queryParameters,
+    Options? options,
     bool validResult = true,
-    BuildContext context,
     bool autoLoading = false}) async {
   if (autoLoading) {
     showLoading();
@@ -116,5 +106,5 @@ Future<Result> requestHttpBin(
     // Dismiss loading
     dismissLoading();
   }
-  return _onValidResult(result, validResult, context);
+  return _onValidResult(result, validResult);
 }
