@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_easy/flutter_easy.dart';
-import 'package:flutter_easy_example/api/tu_chong/tu_chong_model.dart';
+import 'package:flutter_easy_example/api/tu_chong/model/tu_chong_model.dart';
 import 'package:get/get.dart';
 import 'package:photo_view/photo_view_gallery.dart';
 
@@ -11,17 +11,17 @@ class PhotoViewPage extends StatelessWidget {
     return BaseScaffold(
       appBar: BaseAppBar(
         brightness: Brightness.light,
-        title: BaseText(data?.title),
+        title: BaseText(data.title),
       ),
       body: PhotoViewGallery.builder(
         scrollPhysics: const BouncingScrollPhysics(),
         builder: (BuildContext context, int index) {
           return PhotoViewGalleryPageOptions(
             imageProvider:
-                BaseWebImage.provider(data?.images[index].imageURL ?? ""),
+                BaseWebImage.provider(data.imageList?[index]?.imageURL ?? ""),
           );
         },
-        itemCount: data?.imageCount,
+        itemCount: data.imageCount,
         loadingBuilder: (context, event) => Center(
           child: Container(
             width: 20.0,
@@ -29,7 +29,8 @@ class PhotoViewPage extends StatelessWidget {
             child: CircularProgressIndicator(
               value: event == null || event.expectedTotalBytes == null
                   ? 0
-                  : event.cumulativeBytesLoaded / event.expectedTotalBytes,
+                  : event.cumulativeBytesLoaded /
+                      (event.expectedTotalBytes ?? 1),
             ),
           ),
         ),
